apps/ml-nsmf/requirements.txt
# ML-NSMF Requirements

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Machine Learning
# TensorFlow não compatível com Python 3.12 - usando scikit-learn como alternativa
# tensorflow==2.15.0  # Desabilitado - incompatível com Python 3.12
scikit-learn>=1.3.0
# torch>=2.0.0  # Alternativa ao TensorFlow
numpy==1.24.3
pandas==2.1.3

# XAI (Explainable AI)
# shap==0.43.0
# lime==0.2.0.1

# Kafka
kafka-python==2.0.2

# OpenTelemetry
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-exporter-otlp-proto-grpc==1.21.0

# Utilitários
python-dateutil==2.8.2
requests==2.31.0

